# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API para gerenciamento de gastos e informações financeiras",
        "title": "Contabius API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://vkunssec.vercel.app"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/vkunssec/contabius/blob/main/LICENSE"
        },
        "version": "1.0"
    },
    "host": "${HOST}",
    "basePath": "/",
    "paths": {
...
        "/category/update/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Rota para atualizar uma categoria",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Category"
                ],
                "summary": "Rota para atualizar uma categoria",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Dados da categoria",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Categories"
                        }
                    },
                    {
                        "type": "string",
                        "description": "ID da categoria",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/BadRequest"
                        }
                    }
                }
            }
        },
        "/costs": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Rota para retornar os custos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Costs"
                ],
                "summary": "Rota para retornar os custos",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "example": [
                            "678079f6f5080a39a8eedc1e"
                        ],
                        "description": "Ids dos custos a serem retornados",
                        "name": "ids",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Costs"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/BadRequest"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Rota para criar um custo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Costs"
                ],
                "summary": "Rota para criar um custo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Dados do custo",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CostRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/BadRequest"
                        }
                    }
                }
            }
        },
        "/costs/remove/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Rota para deletar um custo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Costs"
                ],
                "summary": "Rota para deletar um custo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID do custo",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/BadRequest"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/BadRequest"
                        }
                    }
                }
            }
        },
        "/costs/update/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Rota para atualizar um custo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Costs"
                ],
                "summary": "Rota para atualizar um custo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Dados do custo",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Costs"
                        }
                    },
                    {
                        "type": "string",
                        "description": "ID do custo",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/BadRequest"
                        }
                    }
                }
            }
        },
...
        "/investments/remove/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Rota para deletar um investimento",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Investments"
                ],
                "summary": "Rota para deletar um investimento",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID do investimento",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/BadRequest"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/BadRequest"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "BadRequest": {
            "description": "Estrutura padrão de resposta de erro de requisição",
            "type": "object",
            "properties": {
                "message": {
                    "description": "Mensagem descritiva",
                    "type": "string",
                    "example": "Erro ao processar a requisição"
                },
                "success": {
                    "description": "Indica se a operação foi bem sucedida",
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "Categories": {
            "type": "object",
            "properties": {
                "category": {
                    "description": "Nome da categoria",
                    "type": "string",
                    "example": "Alimentação"
                },
                "created_at": {
                    "description": "Data de criação",
                    "type": "string",
                    "example": "2025-01-01T00:00:00Z"
                },
                "id": {
                    "description": "ID da categoria",
                    "type": "string",
                    "example": "678079f6f5080a39a8eedc1e"
                },
                "parent": {
                    "description": "ID do pai da categoria",
                    "type": "string",
                    "example": "678079f6f5080a39a8eedc1e"
                },
                "updated_at": {
                    "description": "Data de atualização",
                    "type": "string",
                    "example": "2025-01-01T00:00:00Z"
                }
            }
        },
...
        "CostRequest": {
            "type": "object",
            "required": [
                "amount",
                "category",
                "cost"
            ],
            "properties": {
                "amount": {
                    "description": "Valor do custo",
                    "allOf": [
                        {
                            "$ref": "#/definitions/Money"
                        }
                    ]
                },
                "category": {
                    "description": "ID da categoria",
                    "allOf": [
                        {
                            "$ref": "#/definitions/PartialCategoryRequest"
                        }
                    ]
                },
                "cost": {
                    "description": "Descrição do custo",
                    "type": "string",
                    "example": "aluguel"
                },
                "installments": {
                    "description": "Número de parcelas",
                    "type": "integer",
                    "example": 12
                },
                "methods": {
                    "description": "Método de pagamento",
                    "allOf": [
                        {
                            "$ref": "#/definitions/Methods"
                        }
                    ]
                }
            }
        },
        "Costs": {
            "type": "object",
            "properties": {
                "amount": {
                    "description": "Valor do custo",
                    "allOf": [
                        {
                            "$ref": "#/definitions/Money"
                        }
                    ]
                },
                "category": {
                    "description": "Categoria do custo",
                    "allOf": [
                        {
                            "$ref": "#/definitions/Categories"
                        }
                    ]
                },
                "cost": {
                    "description": "Descrição do custo",
                    "type": "string",
                    "example": "aluguel"
                },
                "created_at": {
                    "description": "Data de criação do custo",
                    "type": "string",
                    "example": "2025-01-01T00:00:00Z"
                },
                "id": {
                    "description": "ID do custo",
                    "type": "string",
                    "example": "678079f6f5080a39a8eedc1e"
                },
                "installments": {
                    "description": "Número de parcelas",
                    "type": "integer",
                    "example": 12
                },
                "methods": {
                    "description": "Método de pagamento",
                    "allOf": [
                        {
                            "$ref": "#/definitions/Methods"
                        }
                    ]
                },
                "updated_at": {
                    "description": "Data de atualização do custo",
                    "type": "string",
                    "example": "2025-01-01T00:00:00Z"
                }
            }
        },
        "Currency": {
            "type": "string",
            "enum": [
                "BRL",
                "USD",
                "EUR"
            ],
            "x-enum-varnames": [
                "CurrencyBRL",
                "CurrencyUSD",
                "CurrencyEUR"
            ]
        },
...
        "Method": {
            "type": "string",
            "enum": [
                "credit",
                "debit",
                "pix",
                "boleto",
                "cash"
            ],
            "x-enum-varnames": [
                "MethodCredit",
                "MethodDebit",
                "MethodPix",
                "MethodBoleto",
                "MethodCash"
            ]
        },
        "MethodId": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                4
            ],
            "x-enum-varnames": [
                "MethodCreditId",
                "MethodDebitId",
                "MethodPixId",
                "MethodBoletoId",
                "MethodCashId"
            ]
        },
        "Methods": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "ID do método de pagamento",
                    "allOf": [
                        {
                            "$ref": "#/definitions/MethodId"
                        }
                    ],
                    "example": 1
                },
                "method": {
                    "description": "Método de pagamento",
                    "allOf": [
                        {
                            "$ref": "#/definitions/Method"
                        }
                    ],
                    "example": "credit"
                }
            }
        },
        "Money": {
            "type": "object",
            "properties": {
                "currency": {
                    "description": "default code BRL",
                    "allOf": [
                        {
                            "$ref": "#/definitions/Currency"
                        }
                    ],
                    "example": "BRL"
                },
                "quantity": {
                    "type": "integer",
                    "example": 1000
                }
            }
        },
...
        "PartialCategoryRequest": {
            "type": "object",
            "required": [
                "category",
                "id"
            ],
            "properties": {
                "category": {
                    "description": "Nome da categoria",
                    "type": "string",
                    "example": "Alimentação"
                },
                "id": {
                    "description": "ID da categoria",
                    "type": "string",
                    "example": "678079f6f5080a39a8eedc1e"
                }
            }
        },
...
        "Response": {
            "description": "Estrutura padrão de resposta",
            "type": "object",
            "properties": {
                "data": {
                    "description": "Dados da resposta"
                },
                "message": {
                    "description": "Mensagem descritiva",
                    "type": "string",
                    "example": "Operação realizada com sucesso"
                },
                "success": {
                    "description": "Indica se a operação foi bem sucedida",
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Bearer token authentication",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}